np-41723: Moved classes to a better package structure
