+++ cristin-organization/src/main/java/no/unit/nva/cristin/organization/fetch/DefaultOrgFetchClientProvider.java
@@ -1,8 +1,9 @@
-package no.unit.nva.cristin.organization;
+package no.unit.nva.cristin.organization.fetch;
 
 import java.util.Map;
 import no.unit.nva.cristin.common.client.IClientProvider;
 import no.unit.nva.cristin.common.client.IFetchApiClient;
+import no.unit.nva.cristin.organization.common.client.CristinOrganizationApiClient;
 import no.unit.nva.model.Organization;
 
 public class DefaultOrgFetchClientProvider
